move to Structure.Serializer
